# api/openapi.yaml
openapi: 3.0.3
info:
  title: Sample API
  description: A short description of API.
  termsOfService: http://example.com/terms/
  contact:
    name: API support
    url: http://www.example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
    description: Local server

tags:
  - name: client
    description: Client
  - name: users
    description: Access to Users
  - name: tweets
    description: Access to Tweets


paths:

  /users/{userId}:
    get:
      tags:
        - users
      summary: Get user by ID.
      description: Returns a single User model
      operationId: GetUserByID
      parameters:
        - name: userId
          in: path
          description: user id
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A single User model
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User'
                example:
                  id: 1
                  name: John Doe

  /tweets/{tweetId}:
    get:
      tags:
        - tweets
      summary: Get tweet by ID.
      description: Returns a single Tweet model
      operationId: GetTweetByID
      parameters:
        - name: tweetId
          in: path
          description: tweet id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single Tweet model
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Tweet'
                example:
                  id: 1
                  text: John Doe
                  user_id: 1

  /oauth/access_token:
    post:
      summary: Authenticate with the admin server.
      description: Used to retrieve all target segments for certain account id.
      operationId: Authenticate
      tags:
        - client
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'

components:
  schemas:
    User:
      type: object
      required:
        - id
      properties:
        id:
          type: integer 
          format: int64
        name:
          type: string

    Tweet:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        user_id:
          type: integer
          format: int64

    AuthenticationRequest:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
        password:
          type: string

    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# security: 
#   # Non-OAuth setting
#   - api_key: []
#   # OAuth setting
#   - users_auth:
#     - write:users
#     - read:users


externalDocs:
  description: Find more info here
  url: https://example.com